core code updates - enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base/trunk@6088 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/server/Server.java
@@ -1421,7 +1421,7 @@ public class Server implements HsqlSocketRequestHandler, Notified {
      * @param cid the connection identifier
      * @param r the request whose description is to be printed
      */
-    final void printRequest(int cid, Result r) {
+    final void printRequest(long cid, Result r) {
 
         if (isSilent()) {
             return;
